pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-auto-commit"
version = "1.1.0"
authors = [
    { name = "Vasanth Adithya", email = "vasanthadithya1234@gmail.com" },
]
description = "A powerful CLI tool for automating GitHub contributions"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "click>=8.1.0",
    "rich>=10.0.0",
    "questionary>=1.10.0",
    "GitPython>=3.1.0",
    "pyfiglet>=0.8.0",
    "schedule>=1.1.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
"Homepage" = "https://github.com/vasanthfeb13/github-auto-commit"
"Bug Tracker" = "https://github.com/vasanthfeb13/github-auto-commit/issues"

[project.scripts]
github-auto-commit = "github_auto_commit.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/github_auto_commit"]

[tool.hatch.build.targets.sdist]
include = [
    "/src/github_auto_commit",
    "/tests",
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
